solve spelling mistakes
