+++ tools/resourcemanager/src/main/java/org/sgdk/resourcemanager/ui/panels/projectexplorer/modals/AddFolderDialog.java
@@ -1,5 +1,6 @@
-package org.sgdk.resourcemanager.ui.panels.proyectexplorer.modals;
+package org.sgdk.resourcemanager.ui.panels.projectexplorer.modals;
 
+import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.GridBagConstraints;
 import java.awt.GridBagLayout;
@@ -7,43 +8,34 @@ import java.awt.Rectangle;
 import java.awt.Toolkit;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
-import java.awt.event.MouseAdapter;
-import java.awt.event.MouseEvent;
 import java.awt.event.WindowAdapter;
 import java.awt.event.WindowEvent;
 import java.io.File;
 
 import javax.swing.JButton;
 import javax.swing.JDialog;
-import javax.swing.JFileChooser;
 import javax.swing.JLabel;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JTextField;
-import javax.swing.filechooser.FileFilter;
 
-import org.apache.commons.lang3.StringUtils;
 import org.sgdk.resourcemanager.entities.SGDKElement;
 import org.sgdk.resourcemanager.entities.SGDKFolder;
-import org.sgdk.resourcemanager.entities.SGDKSprite;
 import org.sgdk.resourcemanager.entities.factory.SGDKEntityFactory;
 import org.sgdk.resourcemanager.ui.ResourceManagerFrame;
 
-public class CreateSpriteDialog extends JDialog{
+public class AddFolderDialog extends JDialog{
 
-	/**
-	 * 
-	 */
 	private static final int minimizeWidth = 340;
 	private static final int minimizeHeight = 220;
 	
-	private JTextField spritePathText = new JTextField();
-	private JFileChooser spritePath = new JFileChooser(System.getProperty("user.home"));
+	private JLabel textError = new JLabel();
+	private JTextField folderPathText = new JTextField();
 	private JButton acceptButon = new JButton("Ok");
 	private static final long serialVersionUID = 1L;
 	
-	public CreateSpriteDialog(ResourceManagerFrame parent, SGDKElement parentNode) {
-		super(parent, "New Sprite");
+	public AddFolderDialog(ResourceManagerFrame parent, SGDKElement parentNode) {
+		super(parent, "Add Folder");
 		parent.setEnabled(false);
 		Dimension screenSize = Toolkit.getDefaultToolkit().getScreenSize();
 		int screenWidth = new Long(Math.round(screenSize.getWidth())).intValue();
@@ -59,52 +51,41 @@ public class CreateSpriteDialog extends JDialog{
 				parent.setEnabled(true);
 			}
 		});
-		
-		spritePath.addChoosableFileFilter(new FileFilter() {			
-			@Override
-			public String getDescription() {
-				return StringUtils.join(SGDKSprite.ValidFormat.values(), ", ");
-			}
-			
-			@Override
-			public boolean accept(File f) {
-				return SGDKSprite.isValidFormat(f.getAbsolutePath()) || f.isDirectory();
-			}
-		});
-		spritePath.setAcceptAllFileFilterUsed(false);
-		
+				
 		setLayout(new GridBagLayout());
 		GridBagConstraints c = new GridBagConstraints();
-			
+		
+		textError.setForeground (Color.red);
+		
 		c.fill = GridBagConstraints.NONE;
 		c.anchor = GridBagConstraints.CENTER;
 		c.weightx = 1d/6d;
 		c.weighty = 1d/2d;		
 		c.gridx = 0;
 		c.gridy = 0;
+		c.gridwidth = 2;
+		c.gridheight = 1;
+		add(textError, c);
+			
+		c.fill = GridBagConstraints.NONE;
+		c.anchor = GridBagConstraints.CENTER;
+		c.weightx = 1d/6d;
+		c.weighty = 1d/2d;		
+		c.gridx = 0;
+		c.gridy = 1;
 		c.gridwidth = 1;
 		c.gridheight = 1;
-		add(new JLabel("Sprite Path: "), c);
+		add(new JLabel("Folder Name: "), c);
 		
 		c.fill = GridBagConstraints.HORIZONTAL;	
 		c.anchor = GridBagConstraints.LINE_START;
 		c.weightx = 5d/6d;
 		c.weighty = 1d/2d;		
 		c.gridx = 1;
-		c.gridy = 0;
+		c.gridy = 1;
 		c.gridwidth = 5;
 		c.gridheight = 1;
-		add(spritePathText, c);		
-		
-		spritePathText.addMouseListener(new MouseAdapter() {
-			@Override
-            public void mouseClicked(MouseEvent e){				
-				int returnVal = spritePath.showDialog(parent, "New Sprite");
-				if (returnVal == JFileChooser.APPROVE_OPTION) {
-					spritePathText.setText(spritePath.getSelectedFile().getAbsolutePath());
-		        }
-            }
-		});
+		add(folderPathText, c);		
 		
 		c.fill = GridBagConstraints.NONE;
 		c.anchor = GridBagConstraints.LINE_END;
@@ -121,30 +102,34 @@ public class CreateSpriteDialog extends JDialog{
 			public void actionPerformed(ActionEvent e) {
 				boolean validForm = true;
 				
-				if(validForm && (spritePathText.getText() == null || spritePathText.getText().isEmpty())) {
+				if(validForm && (folderPathText.getText() == null || folderPathText.getText().isEmpty())) {
 					validForm = false;
 					JPanel panel = new JPanel();
 					JOptionPane.showMessageDialog(panel,
-							 "Invalid Sprite File",
+							 "Invalid Folder Name",
 							 "Error",
 					        JOptionPane.ERROR_MESSAGE);
 				}
 				
 				if (validForm) {
-					SGDKSprite sprite = SGDKEntityFactory.createSGDKSprite(spritePathText.getText(), (SGDKFolder)parentNode);
-					parent.getProyectExplorer().getProyectExplorerTree().addElement(sprite, parentNode);
-					clean();
-					parent.setEnabled(true);
-					setVisible(false);
+					File directory = new File(parentNode.getPath()+File.separator+folderPathText.getText());
+					if(!directory.exists()) {
+						SGDKFolder folder = SGDKEntityFactory.createSGDKFolder(folderPathText.getText(), (SGDKFolder)parentNode);
+						parent.getProjectExplorer().getProjectExplorerTree().addElement(folder, parentNode);
+						clean();
+						parent.setEnabled(true);
+						setVisible(false);
+					}else {
+						textError.setText("Carpeta ya existe");
+					}
 				}
 			}
-		});
-		
+		});		
 		setVisible(true);
 	}
 
 	protected void clean() {
-		spritePathText.setText("");
+		folderPathText.setText("");
+		textError.setText("");
 	}
-
 }
