+++ tools/resourcemanager/src/main/java/org/sgdk/resourcemanager/ui/panels/projectexplorer/modals/CreateBackgroundDialog.java
@@ -1,4 +1,4 @@
-package org.sgdk.resourcemanager.ui.panels.proyectexplorer.modals;
+package org.sgdk.resourcemanager.ui.panels.projectexplorer.modals;
 
 import java.awt.Dimension;
 import java.awt.GridBagConstraints;
@@ -23,24 +23,28 @@ import javax.swing.JTextField;
 import javax.swing.filechooser.FileFilter;
 
 import org.apache.commons.lang3.StringUtils;
+import org.sgdk.resourcemanager.entities.SGDKBackground;
 import org.sgdk.resourcemanager.entities.SGDKElement;
-import org.sgdk.resourcemanager.entities.SGDKFXSound;
 import org.sgdk.resourcemanager.entities.SGDKFolder;
 import org.sgdk.resourcemanager.entities.factory.SGDKEntityFactory;
 import org.sgdk.resourcemanager.ui.ResourceManagerFrame;
 
-public class CreateSFXDialog extends JDialog{
+public class CreateBackgroundDialog extends JDialog{
+
+	/**
+	 * 
+	 */
 	
 	private static final int minimizeWidth = 340;
 	private static final int minimizeHeight = 220;
 	
-	private JTextField fxSoundPathText = new JTextField();
-	private JFileChooser fxSoundPath = new JFileChooser(System.getProperty("user.home"));
+	private JTextField backgroundPathText = new JTextField();
+	private JFileChooser backgroundPath = new JFileChooser(System.getProperty("user.home"));
 	private JButton acceptButon = new JButton("Ok");
-	private static final long serialVersionUID = 1L;
+	private static final long serialVersionUID = 1L;	
 	
-	public CreateSFXDialog(ResourceManagerFrame parent, SGDKElement parentNode) {
-		super(parent, "New FX Sound");
+	public CreateBackgroundDialog(ResourceManagerFrame parent, SGDKElement parentNode) {
+		super(parent, "New Background");
 		parent.setEnabled(false);
 		Dimension screenSize = Toolkit.getDefaultToolkit().getScreenSize();
 		int screenWidth = new Long(Math.round(screenSize.getWidth())).intValue();
@@ -57,18 +61,18 @@ public class CreateSFXDialog extends JDialog{
 			}
 		});
 		
-		fxSoundPath.addChoosableFileFilter(new FileFilter() {			
+		backgroundPath.addChoosableFileFilter(new FileFilter() {			
 			@Override
 			public String getDescription() {
-				return StringUtils.join(SGDKFXSound.ValidFormat.values(), ", ");
+				return StringUtils.join(SGDKBackground.ValidFormat.values(), ", ");
 			}
 			
 			@Override
 			public boolean accept(File f) {
-				return SGDKFXSound.isValidFormat(f.getAbsolutePath()) || f.isDirectory();
+				return SGDKBackground.isValidFormat(f.getAbsolutePath()) || f.isDirectory();
 			}
 		});
-		fxSoundPath.setAcceptAllFileFilterUsed(false);
+		backgroundPath.setAcceptAllFileFilterUsed(false);
 		
 		setLayout(new GridBagLayout());
 		GridBagConstraints c = new GridBagConstraints();
@@ -81,7 +85,7 @@ public class CreateSFXDialog extends JDialog{
 		c.gridy = 0;
 		c.gridwidth = 1;
 		c.gridheight = 1;
-		add(new JLabel("FX Sound Path: "), c);
+		add(new JLabel("Background Path: "), c);
 		
 		c.fill = GridBagConstraints.HORIZONTAL;	
 		c.anchor = GridBagConstraints.LINE_START;
@@ -91,14 +95,14 @@ public class CreateSFXDialog extends JDialog{
 		c.gridy = 0;
 		c.gridwidth = 5;
 		c.gridheight = 1;
-		add(fxSoundPathText, c);		
+		add(backgroundPathText, c);		
 		
-		fxSoundPathText.addMouseListener(new MouseAdapter() {
+		backgroundPathText.addMouseListener(new MouseAdapter() {
 			@Override
             public void mouseClicked(MouseEvent e){				
-				int returnVal = fxSoundPath.showDialog(parent, "New FX Sound");
+				int returnVal = backgroundPath.showDialog(parent, "New Background");
 				if (returnVal == JFileChooser.APPROVE_OPTION) {
-					fxSoundPathText.setText(fxSoundPath.getSelectedFile().getAbsolutePath());
+					backgroundPathText.setText(backgroundPath.getSelectedFile().getAbsolutePath());
 		        }
             }
 		});
@@ -118,21 +122,21 @@ public class CreateSFXDialog extends JDialog{
 			public void actionPerformed(ActionEvent e) {
 				boolean validForm = true;
 				
-				if(validForm && (fxSoundPathText.getText() == null || fxSoundPathText.getText().isEmpty())) {
+				if(validForm && (backgroundPathText.getText() == null || backgroundPathText.getText().isEmpty())) {
 					validForm = false;
 					JPanel panel = new JPanel();
 					JOptionPane.showMessageDialog(panel,
-							 "Invalid FX Sound File",
+							 "Invalid Background File",
 							 "Error",
 					        JOptionPane.ERROR_MESSAGE);
 				}
 				
 				if (validForm) {
-					SGDKFXSound fxSound = SGDKEntityFactory.createSGDKFXSound(fxSoundPathText.getText(), (SGDKFolder)parentNode);
-					parent.getProyectExplorer().getProyectExplorerTree().addElement(fxSound,parentNode);
+					SGDKBackground background = SGDKEntityFactory.createSGDKBackground(backgroundPathText.getText(), (SGDKFolder)parentNode);
+					parent.getProjectExplorer().getProjectExplorerTree().addElement(background, parentNode);
 					clean();
 					parent.setEnabled(true);
-					setVisible(false);
+					setVisible(false);					
 				}
 			}
 		});
@@ -141,6 +145,7 @@ public class CreateSFXDialog extends JDialog{
 	}
 
 	protected void clean() {
-		fxSoundPathText.setText("");
+		backgroundPathText.setText("");
 	}
+
 }
